requirements.txt
openai==0.26.4
pyyaml
psycopg2